refactor(core): relocate upstream C sources and headers
Motivation: The original top-level source and include directories made
the project-owned runtime difficult to distinguish from build tooling,
generated outputs, and deployment concerns. The checked-in packer
binary also obscured which source files were authoritative.

Details: Move runtime headers and implementations into core/inc and
core/src, place the host packer interface and implementation under
tools/packer, update affected include paths, and remove the obsolete
packer executable from version control.

Impact: Establishes explicit ownership boundaries for the WebAssembly
runtime and host tooling, so later build, analysis, and review changes
can target stable paths without relying on the legacy repository
structure.

+++ core/inc/ui.h
@@ -94,7 +94,7 @@ void ui_sync_url(const char *path);
 /*
  * ui_get_current_hash - Returns the current URL hash into the provided buffer.
  */
-void ui_get_current_hash(char *buf, int max_len);
+void ui_get_current_hash(char *buf, size_t max_len);
 
 /*
  * add_bar - Appends a segmented graph bar to the UI.
